refactored useIsMobile hook
