--- frontend/src/hooks/PageContext.tsx
@@ -1,11 +1,0 @@
-import { createContext, useContext } from "react";
-
-const PageContext = createContext(() => {
-  const isMobile = window.matchMedia("(max-width: 768px)").matches
-
-  return {
-    isMobile
-  }
-})
-
-export default PageContext
