+++ frontend/src/pages/LandingPage/components/HeroImage.tsx
@@ -1,11 +1,11 @@
 import React, { useContext } from "react"
-import PageContext from "../../../hooks/PageContext"
+import useIsMobile from "../../../hooks/useIsMobile"
 import heroImg from '../../../assets/images/hero-image-6.jpg'
 import Button from "../../../components/Button"
 import {ReactComponent as EngagmentRing} from '../../../assets/icons/engagement-ring.svg'
 
 const HeroImage = () => {
-  const { isMobile } = useContext(PageContext)()
+  const { isMobile } = useIsMobile()
 
   return (
     <div style={{
