+++ frontend/src/pages/LandingPage/index.tsx
@@ -1,9 +1,5 @@
 import React from 'react';
-import Carousel from './components/Carousel'
 import './style.scss'
-import PageContainer from '../../components/PageContainer';
-import Button from '../../components/Button';
-import {ReactComponent as EngagmentRing} from '../../assets/icons/engagement-ring.svg'
 import HeroImage from './components/HeroImage';
 
 const LandingPage = () => {
